+++ frontend/src/app/favorites/page.tsx
@@ -10,20 +10,20 @@ export default async function Favorites() {
     <div className="mx-auto max-w-lg">
       {favoritesResult.success
         ? (
-          <FavoritesList
-            initialData={favoritesResult.data}
-            initialMeta={favoritesResult.meta}
-          />
-        )
+            <FavoritesList
+              initialData={favoritesResult.data}
+              initialMeta={favoritesResult.meta}
+            />
+          )
         : (
-          <div className="my-10 text-center">
-            <h2 className="mb-2">お気に入りを読み込めませんでした。</h2>
-            <p className="mb-6">時間をあけてから、再度お試しください。</p>
-            <Button asChild size="lg">
-              <Link href="/">トップページに戻る</Link>
-            </Button>
-          </div>
-        )}
+            <div className="my-10 text-center">
+              <h2 className="mb-2">お気に入りを読み込めませんでした。</h2>
+              <p className="mb-6">時間をあけてから、再度お試しください。</p>
+              <Button asChild size="lg">
+                <Link href="/">トップページに戻る</Link>
+              </Button>
+            </div>
+          )}
     </div>
   )
 }
