+++ frontend/src/components/features/account/auth/AuthDrawer.tsx
@@ -3,21 +3,13 @@
 import { useAtom } from 'jotai'
 import Image from 'next/image'
 import { useState } from 'react'
+import { Button } from '~/components/ui/button'
+import { Drawer, DrawerContent, DrawerDescription, DrawerHeader, DrawerTitle, DrawerTrigger } from '~/components/ui/drawer'
 import { AUTH_PROVIDERS, dynamicPaths } from '~/constants'
-import logoMark from '~/public/logo_mark.webp'
 import { authModalOpenAtom } from '~/state/auth-modal-open.atom'
 import { apiHref } from '~/utils/api-url'
-import { Button } from '../button'
-import LoginForm from '../forms/LoginForm'
-import SignUpForm from '../forms/SignUpForm'
-import {
-  Drawer,
-  DrawerContent,
-  DrawerDescription,
-  DrawerHeader,
-  DrawerTitle,
-  DrawerTrigger,
-} from '../drawer'
+import LoginForm from './forms/LoginForm'
+import SignUpForm from './forms/SignUpForm'
 
 export function AuthDrawer() {
   const [open, setOpen] = useAtom(authModalOpenAtom)
@@ -44,7 +36,7 @@ export function AuthDrawer() {
             <div className="flex items-center gap-2 py-1">
               <Image
                 alt="mark"
-                src={logoMark}
+                src="/logo_mark.webp"
                 width={24}
                 height={24}
               />
