+++ frontend/src/components/features/station/search/StationSearchForm.tsx
@@ -6,14 +6,14 @@ import { useRouter } from 'next/navigation'
 import { useFieldArray, useForm } from 'react-hook-form'
 import { toast } from 'sonner'
 import { Button } from '~/components/ui/button'
+import { Form, FormControl, FormField, FormItem } from '~/components/ui/form'
 import { useMidpointMutation } from '~/hooks/useMidpointMutation'
 import { logger } from '~/lib/logger'
 import { stationSearchSchema } from '~/schemas/station-search.schema'
+import { RemoveFormButton } from './/buttons/RemoveFormButton'
+import { AddFormButton } from './buttons/AddFormButton'
+import { ResetFormButton } from './buttons/ResetFormButton'
 import StationAutocomplete from './StationAutocomplete'
-import { AddFormButton } from '../../../ui/buttons/AddFormButton'
-import { RemoveFormButton } from '../../../ui/buttons/RemoveFormButton'
-import { ResetFormButton } from '../../../ui/buttons/ResetFormButton'
-import { Form, FormControl, FormField, FormItem } from './Form'
 
 const MAX_AREA_FIELDS = 6
 const MAX_REQUIRED_FIELDS = 2
@@ -56,8 +56,8 @@ export default function StationSearchForm() {
 
           const isSameStationIds
             = sortedPrevIds.length === sortedStationIds.length
-            && sortedPrevIds.every((prevStationId, i) =>
-              prevStationId === sortedStationIds[i])
+              && sortedPrevIds.every((prevStationId, i) =>
+                prevStationId === sortedStationIds[i])
 
           if (!isSameStationIds) {
             sessionStorage.removeItem('pendingSearchHistoryId')
