+++ frontend/src/components/features/station/search/buttons/AddFormButton.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import type { FieldArrayWithId, UseFieldArrayAppend } from 'react-hook-form'
 import type { AreaFormValues } from '~/schemas/station-search.schema'
 import { LuPlus } from 'react-icons/lu'
